c-opts.c (missing_arg): Use cl_options[opt_index].opt_code instead of just opt_index as switch expression.
	* c-opts.c (missing_arg): Use cl_options[opt_index].opt_code
	instead of just opt_index as switch expression.

	* calls.c (store_one_arg): Change type of 'excess_align'
	to unsigned int.

	* profile.c (output_gcov_string): Change type of 'temp'
	to size_t.

From-SVN: r58381

+++ gcc/profile.c
@@ -221,7 +221,7 @@ output_gcov_string (string, delimiter)
      const char *string;
      long delimiter;
 {
-  long temp;
+  size_t temp;
 
   /* Write a delimiter to indicate that a file name follows.  */
   __write_long (delimiter, bb_file, 4);
